Implement related elements.
Refactored as a single list of type references, similar to
history.TypeRef. A single list is possible, because the type category
can be used to filter different types of elements.

Only members can point to other elements through their fields. So the
Related field is a list of TypeRef values. These each refer to a
member, but also drill all the way down to the specific field.
Additional information is included to allow filtering. Enough
information is included to reconstruct the reference without lookups.
TypeRefs are sorted as is reasonable for display.