Do not escape file names.
Hugo is not correctly handling links to URL-escaped file names. For now,
generate file names as-is.

This means that a file named "<<<ROOT>>>.md" will be generated, which
contains characters not allowed on Windows. I don't care about Windows,
so this is fine.

This also means that the page generated by Hugo will be "ROOT.html".
While there may be no class named "ROOT" or "Root" right now, that
doesn't stop Roblox from adding it in the future. Hugo wont conflict
with a "Root" class, but it will with a "ROOT" class. The latter seems
extremely unlikely ever be added, so this is fine. "Root" would
conflict on Windows, but once again, I don't care.

Additionally, this will break any existing links to the <<<ROOT>>> page,
which originally had a URL-escaped name. This may or may not be
resolvable with aliases.

Restoring this behavior would involve submitting an issue to Hugo
indicating that links are not sanitized correctly.