Implement actions.
Settings cause entities with certain statuses to become visible or
hidden. These changes can cause certain information on the page to
become irrelevant or incorrect. For example, the counter that counts
the number of members will be incorrect if deprecated members are
hidden. Or, if all related classes are hidden, then the entire section
should also be hidden.

Actions update elements based on visibility of other elements. A target
element is selected, having a predefined function applied to it. The
function receives a subject element, and a selector that matches
elements to be counted within it.

- The Count function sets the content of the target to the number of
  visible selected elements within the subject.
- The HideIfZero function hides the target if the number of visible
  selected elements within the subject is zero.