Fix icons not respecting theme.
Icons used media queries to select an image based on
preferred-color-scheme. Unfortunately, because theme overriding is
handled by CSS, there was no way to override the icon.

Icons are now rendered by CSS. This enables selecting icons based on the
overridden theme, as well as preferred-color-scheme when the theme is
set to Auto.

An icon image is set according to the --light and --dark variables,
which are expected to be any value receivable by the background-image
property. If one variable is unspecified, then the other is used. If
neither are specified, then `none` is used.